requirements.txt
discord.py
tweepy
python-dotenv